app_rpt:  Channel and GPIO initialization for PTT
In chan_simpleusb and chan_usbradio, updated the code that manipulates the
PTT GPIO pin.  Specifically, make sure that we initialize the pin output
value when configured with "invertptt = yes".  Also ensure that changes
to the PTT pin state does not affect the state for any other pins (e.g.
the "clip" LED pin).