app_rpt:  Channel initialization for PTT (#426)
In chan_simpleusb and chan_usbradio, updated the code that manipulates
the PTT GPIO pin.  Specifically, make sure that we initialize the pin output
value when configured with "invertptt = yes".  Also ensure that changes
to the PTT pin state does not affect the state for any other pins (e.g.
the "clip" LED pin).

Now that we are correctly initializing the GPIO pins at startup the code
to toggle PTT (that wasn't always processed) is no longer needed.

Co-authored-by: Allan Nathanson <42244061+Allan-N@users.noreply.github.com>