res_usbradio: Add ability to specify allowed vid:pid (#428)
This updates res_usbradio to allow users to specify USB device descriptor(s) that can be used for interface devices.  There are compatible chips that are not supported in the code.  This will allow users to easily add the compatible chip(s).

This adds res_usbradio.conf for configuration.  A sample configuration file was also added.

This closes #359.