CI: Don't restart Asterisk in the middle of the build.
It seems that /etc/init.d/asterisk restart is now returning
nonzero, causing the build to abort. This action isn't
actually even necessary, so we could remove it, or safely
continue the build regardless of its return value.

Also fix an issue with rpt_install.sh so the path for
the audio files is correct in the CI.