app_rpt: Don't hard hangup channels currently executing PBX.
process_link_channels is called by the rpt thread for a repeater,
and in turn calls remote_hangup_helper in 1 of 2 places. (This
arrangement has existed since the refactoring in commit
3834be01031a4098a14b2c36f99f99a26223f396.)

remote_hangup_helper sets l->chan to NULL, but if l->chan happens
to be the channel passed to rpt_exec, then this interferes with
the execution of Rpt(). This is possible since rpt_exec calls
rpt_link_add, which imparts l into the list, at which point
other threads could access it. Normally, this is fine, since
Rpt() immediately returns and terminates the PBX on the channel
(formerly "keepalive"), but if rpt_exec is still running at this
point, then if l->chan is NULL, we can't even use chan because
ast_hangup (which is a hard hangup) has already been called on it.
It is, in essence, a classic race condition, triggered only when
an incoming connection is immediately disconnected by the node
thread. A hard hangup at this point in time is invalid because
a PBX thread is still using the channel.

This is, for example, triggered by the newkeytimer expiration
issue, but could be triggered by other things as well.

Fortunately, there is a simple check we can do. As long as
the channel is executing Rpt(), it has a PBX, until the very end.
Therefore, we should simply avoid calling ast_hangup on any
channel with a PBX, as that will always result in illegal
behavior. In the PBX thread, after imparting the channel into
the list of links, we also take care to lock the channel and
check if l->chan has been set NULL or a softhangup has been
queued on the channel. We also use the chan variable directly
in case l->chan has been set NULL.

As part of this change:

* Simplify return paths in process_link_channels. Previously, we
  would lock myrpt->lock, break out of the loop, unlock,
  and then return immediately. It is simpler and faster to skip
  the unnecessary lock/unlock altogether and return directly
  where appropriate. remote_hangup_helper also now returns without
  the lock being held, since it isn't needed, instead of locking
  before returning (which was confusing since the lock wasn't
  held when entering the function, only when leaving it).
* We now assert the channel is not NULL in rpt_sendtext.

Resolves: #422