app_gps: Code updates and documentation
This updates app_gps with the following changes:

Update the code to use the asterisk coding guidelines. Added documentation to all functions.
Added documentation to gps.conf.
Added functions: GPS_READ and APRS_SENDTT.
Added the cli command gps status.
Added the ability to select the primary or alternate symbol table. Replaced depreciated ast_gethostbyname.
Addressed issues with module unload not completing timely.

GPS_READ() function replaces the need to create the gps.dat file stored in the /tmp directory. The application previously wrote to the disk once per second, which included a system call to rename a temporary file.

app_rpt and chan_echolink have been updated to use GPS_READ to retrieve the gps data, instead of reading a data file.

APRS_SENDTT(stanza, overlay) = callsign, function replaces the pipe that was used by app_rpt to send APRStt requests to app_gps.

APSTAR is now used as the tocall for both APRS an APRStt posts. This is being registered as the tocall for AllStar Link.

If elevation is available, it is included in the APRS post.

The cli command 'gps status' will allow the user to check the status of the attached GPS device. It tells the user if the GPS device is connected, if it is locked to the satellites, and the current and default position.

The unlocked warning message is now only printed when it first occurs. The message is turned back on when we receive a lock on the satellite.

The gps.conf file now has an entry for icontable. The default is '' primary or '/' for the alternate symbol table.

For APRStt to work, the aprstt entry in app_rpt needs to be enabled. That option has been added to rtp.conf and commented out.

There were a number of refinements to the code, along with fixes for memory leaks.