app_rpt: Don't emit warnings about <sys/io.h> being unavailable
While some may argue that a few known/understood compilation warnings are
OK we get into trouble when stop looking for "new" warnings in build output
that includes a number of "old" ones. The simple solution is to banish ALL
warnings from the project and then treat any warnings as errors.

This PR addresses the `sys.io is not available on this architecture` warnings
reported on `arm64` builds.