app_rpt: Eliminate START_DELAY and use per-node logic instead.
Currently, Rpt() rejects all calls received within the first two
seconds after the repeater's thread is launched, to give the
thread time to start. This is an arbitrary wait time that
unnecessarily delays node connects in tests and doesn't guarantee
anything is actually started properly. Instead of using an arbitary
number for all nodes, look at whether the specific node is actually
ready or not yet.