tests: Add initial test suite test for fast disconnect.
As part of work on troubleshooting the issues raised in #459
and #460, add a test to try to capture this condition.
The test doesn't reliably capture that issue (it passes
most of the time), but it should continue to pass if there
is a fix.

Since this is our first test, some changes have been made
to rpt_install.sh to allow patching the tests in if the
Asterisk test suite is present. The CI also runs the tests
now.

Note: This is a test suite (black box) test. Currently, there
aren't any unit tests (white box) tests.

Part of #20