app_rpt.c: fix non channel type "irlp"
Original code was looking for strings in the channel name such that IRLP and Echolink was in the name.  As written, gain applied to fs was not making it to the channel.
fs is a dup of f, then gain is applied or not. There's no reason not to just use f for all ast_write() and eliminate fs and the dup.