app_rpt: Properly work around removal of AST_PBX_KEEPALIVE.
In the original refactor of app_rpt for current versions of Asterisk,
we hacked around the fact that AST_PBX_KEEPALIVE was removed back
in https://github.com/asterisk/asterisk/commit/50a25ac8474d7900ba59a68ed4fd942074082435.
At the time, there was no correct way to emulate what we were trying to do,
as ast_pbx_run_args could only be used when the caller knew the channel
wouldn't hang up in advance, and the PBX doesn't do that.

The hack at the time to compile was simply to call pthread_exit() to
abort the PBX thread and prevent it from hanging up the channel as it
would have when it cleaned up after dialplan finished. This was far
from ideal as it was an unclean and improper PBX termination.

Since the PBX args interface doesn't allow setting an arg after the PBX
has been started, we instead masquerade the channel so that the original
channel is killed off by the PBX thread (which is fine) and the masqueraded
channel is what gets used by the repeater threads.

Resolves: #15