simpleusb, usbradio: skip categories with no "devstr"
While testing some new menu / "customization" changes I found that the
initial "active" device configuration was set to a category that did not
have a "devstr" variable setting.  In this case, the category was intended
to augment a templated configuration.  To ensure that these categories are
not being considered we now look for the presence of a "devstr" variable.