app_rpt: If we are unkeyed, send an AST_FRAME_CNG. (#596)
If we are unkeyed, send an AST_FRAME_CNG to stop the jitter buffer from interpolating missing frames.

Specifically when a loud noise (like a squelch burst) occurs in the last frame, the jitter buffer and the codec try to interpolate the missing frames for up to 10 frames.  Sending an AST_FRAME_CNG frame triggers the interpolator to not interpolate