app_rpt: add "archivedatefmt" variable for archive date string
Setting the "archivedir" variable enables a simple log and audio recording
functionality.  Our documentation claimed that the file names would reflect
the date and time down to the 1/100 of a second.  In reality, the names
were limited to 1 second.

This change adds a new "archivedatefmt" variable allowing one to specify
the date/time formatting.  The default format string remains the same but
one can now add additional sub-second digits.