app_rpt: Update duplex mode 3 to pass audio to audioarchive. (#603)
Mode 3 is intended to be a full duplex with local audio muted.  This WAS accomplished via conference settings but that prevented the local audio from reaching the monstream channel.
Converting the conference settings to match modes 2 and 4 (full duplex) and simply muting the frames headed to the txchannel allows for monstream to catch the audio and prevent the audio from feeding back on the local speaker.
Resolves issue archivedir recording (mostly) silence to WAV file when duplex=3
Resolves #600