chan_echolink: fix crash associated with incoming connection
An incoming EchoLink connection to a node already associated with another
node with [many] linked nodes was triggering a crash.  The prime suspect
for the crash was the el_text() function that was iterating over the linked
nodes and generating text to be passed to connected EchoLink nodes.  This
code was hard to read, had some fixed-sized buffers, and did a lot of
memory copying.  The code has been reworked and appears to have resolved
the crashes.

A handful of other changes, mostly for readability, are included.