app_rpt: add "archiveaudio" variable to control audio capture
Adding a new (to ASL3) "archiveaudio" variable that can be used to control
whether the both the simple activity log and audio recordings are included
in the archive directory.  The variable defaults to "yes" (capture audio)
but if set to "no" then only the log will be created (the audio recordings
will not be saved).

This added functionality might be used by those who want to keep track of
node activity without the added expense (and storage) associated with
saving the audio content.