simpleusb: Fix SimpleUSB tuning
One of the changes made with https://github.com/AllStarLink/app_rpt/pull/605
combined the "Carrier From" and "CTCSS From" enum values used in the
SimpleUSB and USBRadio channel drivers (and tune menus).  Unfortunately,
the SimpleUSB code that mapped between the configuration file string values
and the [new] enum values was not updated.