app_rpt: ignore !IAXKEY! message
The IAXRpt Windows client generates an "old" `!IAXKEY!` message that was
essentially not being used in `app_rpt`.  We were not aware of this usage
and yanked the code that handled that message in [PR #588](https://github.com/AllStarLink/app_rpt/pull/588).

It turns out that not handling the message resulted in a WARNING message
being logged.  This PR restores just enough of the [old] code to ensure
that the warning is no longer issued.