app_rpt: fix [echolink] choppy audio (#723)
First off, a HUGE thank you to Chuck/AI7SY for his help in identifying the PR that resulted in the choppy audio.

The changes that were made for #538 had a few places where we updated the `struct rpt` frame queue (lastf1, lastf2) but should have updated the `struct rpt_link` frame queue.  This change restores the expected behavior.