Remove invalid options `txmixa` and `txmixb`. Update definitions.
`txmixa` and `txmixb` are not valid options for `chan_simpleusb`.
Remove them from the default .conf file. Update descriptions
of `rxondelay` and `txoffdelay` for better clarity of what they
do, and how they interact with each other. Correct description
that default option for `eeprom` is `1` (not `0`). Add link
to ASL3 Manual page for `simpleusb.conf`.

Closes: #746