app_rpt.c: Warn when using pipe delimiter in dialplan instead of comma.
The pipe (|) delimiter was previously used in OLD Asterisk. The
current delimiter is the comma (,). We currently support either, but
the plan is to stop supporting the pipe at sometime in the future.

This change warns users that there is a pipe in their dialplan (in
extensions.conf) and reminds them to change it.

UpgradeNote: Users should update calls to Rpt to use the comma delimiter
to prevent future errors.