chan_voter: Detect and handle buflen too small with mix mode clients (#867)
* chan_voter: Detect and handle buflen too small with mix mode clients

Re-submitting an earlier fix that strictly deals with this issue.

This fix properly detects when a client tries connecting in mix mode,
and the buflen in voter.conf is too small (previously known as an
"out of bounds" error). Now, we will log an error telling the user
to fix buflen in voter.conf, and block the mix mode clients from
connecting (they can't send audio to the channel when they are misconfigured
anyways).

Resolves #847

* Fix indents.

* Incorporate CodeRabbit suggestions.

* Get rid of magic number and compare against FRAME_SIZE.

* Clang formatting and remove whitespace.