app_rpt: address race setting `myrpt->calldigittimer` (#864)
When calling a number from the command line with a partial match ex: `rpt fun 1999 *61234` there is a race between the extension check setting `myrpt->calldigittimer = 1` and the `rpt_call()` thread setting `myrpt->calldigittimer = 0`.  This prevents a partial match extension from dialing and always ends in a patch timeout.
The idea behind `myrpt->calldigittimer = 0` is to prevent digit timing until the first digit is received.  If a user dials *61, the patch timer should start, but the inter digit timers should wait until the first extension digit is dialed.