chan_voter: Remove nulaw audio processing not implemented in clients.
The VOTER/RTCM can only send ulaw or ADPCM audio. The "Nulaw" code in
`chan_voter` can never be used, and is therefore "dead code". The
Nulaw implementation is a low quality audio encoding, and is of no
practical use. Due to the code space restrictions in the VOTER/RTCM,
there is no way to implement it on those devices.

Note that the DSP Low Pass filter was only ever used for Nulaw audio
processing, hence why it too has been removed.

Resolves #872